+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 Changelog
 =========
+## Unreleased
+### Fixed
+- CP notifications were not being displayed after field actions occured
 
 ## 1.0.0-beta.1 - 2018-04-26
 ### Fixed